\name{AromaSort}
\alias{AromaSort}
\docType{data}
\title{
  Aroma sorting data
}
\description{
  Partitions of 16 aromas by 31 subjects (free sorting task)
}
\usage{data(AromaSort)}
\format{
  A data frame with 16 observations (aromas) and 31 variables (subjects).
}
\details{
  List of stimuli : Lemon, Grapefruit, Pineapple, Pear, Honey, Butter, Grilledbread, Grilledhazelnut, Strawberry, Raspberry, Cherry, Blackcurrant, Greenpepper, Smoked, Pepper, Licorice.
}
\references{
  Ph. Courcoux, P. Faye, E.M. Qannari (2014) Determination of the consensus partition and cluster analysis of subjects in a free sorting task experiment. Food Quality and Preference 32, 107-112

}
\examples{
data(AromaSort)
}
